provide option to remove previous configuration
When doing some UPI testing, I found myself tearing down all the infra
and using the playbook to repopulate things.  This was kind of hit or
miss because I would forget to update a file or a MAC address.  This
gives users the ability to nuke and pave a lot of the config files
laid down by the playbook before generating new files.